Rename files to ts and update imports
