+++ lib/clamp.ts
