+++ lib/dec2frac.ts
