+++ lib/numberProps.ts
