+++ lib/options.ts
