+++ lib/pad.ts
