+++ lib/runPart.ts
@@ -1,17 +1,17 @@
-import { round } from './round.js';
-import { clamp } from './clamp.js';
-import { dec2frac } from './dec2frac.js';
-import { general } from './general.js';
-import { toYMD } from './toYMD.js';
-import { defaultLocale } from './locale.js';
+import { round } from './round.ts';
+import { clamp } from './clamp.ts';
+import { dec2frac } from './dec2frac.ts';
+import { general } from './general.ts';
+import { toYMD } from './toYMD.ts';
+import { defaultLocale } from './locale.ts';
 import {
   u_DSEC, u_CSEC, u_MSEC,
   EPOCH_1317,
   MIN_S_DATE, MAX_S_DATE,
   MIN_L_DATE, MAX_L_DATE
-} from './constants.js';
-import { pad } from './pad.js';
-import { getExponent, getSignificand } from './numberProps.js';
+} from './constants.ts';
+import { pad } from './pad.ts';
+import { getExponent, getSignificand } from './numberProps.ts';
 
 const DAYSIZE = 86400;
