+++ test/fragmented-numbers.spec.ts
@@ -1,5 +1,5 @@
 import { expect, test } from 'vitest';
-import { format } from '../lib/index.js';
+import { format } from '../lib/index.ts';
 
 test('Fragmented number parts', () => {
   // exps with fragments
