+++ test/ssf-valid.spec.ts
@@ -1,5 +1,5 @@
 import { expect, test } from 'vitest';
-import { format } from '../lib/index.js';
+import { format } from '../lib/index.ts';
 
 test('Validity tests', () => {
   expect(() => format('" Excellent"', 1)).not.toThrow();
